Refactor: 피드백 반영, 구매, 판매 (Purchase, Sale)로 단일화
