--- src/main/java/com/uniqueAuction/domain/trade/entity/PurchaseType.java
@@ -1,6 +1,0 @@
-package com.uniqueAuction.domain.trade.entity;
-
-public enum PurchaseType {
-    PURCHASE_NOW, /* 즉시 구매 */
-    PURCHASE_BID, /* 구매 입찰 */
-}
