--- src/main/java/com/uniqueAuction/domain/trade/entity/SaleType.java
@@ -1,6 +1,0 @@
-package com.uniqueAuction.domain.trade.entity;
-
-public enum SaleType {
-    SALE_NOW, /* 즉시 판매 */
-    SALE_BID /* 판매 입찰 */
-}
