+++ src/main/java/com/uniqueAuction/domain/trade/entity/TradeStatus.java
@@ -2,6 +2,5 @@ package com.uniqueAuction.domain.trade.entity;
 
 public enum TradeStatus {
     BID_PROGRESS, /* 입찰 진행중 */
-    BID_COMPLETE, /* 입찰 체결 */
-    TRADE_COMPLETE /* 거래 완료*/
+    BID_COMPLETE /* 입찰 체결 */
 }
