+++ src/main/java/com/uniqueAuction/web/trade/controller/PurchaseController.java
@@ -2,6 +2,7 @@ package com.uniqueAuction.web.trade.controller;
 
 import java.util.Objects;
 
+import com.uniqueAuction.web.trade.request.PurchaseRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
@@ -12,9 +13,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.uniqueAuction.domain.trade.service.PurchaseService;
 import com.uniqueAuction.web.response.CommonResponse;
-import com.uniqueAuction.web.trade.request.PurchaseBidRequest;
-import com.uniqueAuction.web.trade.request.PurchaseNowRequest;
-
 import lombok.RequiredArgsConstructor;
 
 @RestController
@@ -23,25 +21,14 @@ public class PurchaseController {
 
 	private final PurchaseService purchaseService;
 
-	@PostMapping("/purchaseBid")
+	@PostMapping("/purchase")
 	@ResponseStatus(HttpStatus.CREATED)
-	public CommonResponse savePurchaseBid(@RequestBody @Validated PurchaseBidRequest purchaseBidRequest,
+	public CommonResponse<?> savePurchaseNow(@RequestBody @Validated PurchaseRequest purchaseRequest,
 		BindingResult result) {
 		if (result.hasErrors()) {
 			throw new RuntimeException(Objects.requireNonNull(result.getFieldError()).getDefaultMessage());
 		}
-		purchaseService.savePurchaseBid(purchaseBidRequest.toEntity());
-		return CommonResponse.success();
-	}
-
-	@PostMapping("/purchaseNow")
-	@ResponseStatus(HttpStatus.CREATED)
-	public CommonResponse savePurchaseNow(@RequestBody @Validated PurchaseNowRequest purchaseNowRequest,
-		BindingResult result) {
-		if (result.hasErrors()) {
-			throw new RuntimeException(Objects.requireNonNull(result.getFieldError()).getDefaultMessage());
-		}
-		purchaseService.savePurchaseNow(purchaseNowRequest.toEntity());
+		purchaseService.savePurchase(purchaseRequest.toEntity());
 		return CommonResponse.success();
 	}
 }
