+++ src/main/java/com/uniqueAuction/web/trade/controller/SaleController.java
@@ -2,6 +2,7 @@ package com.uniqueAuction.web.trade.controller;
 
 import java.util.Objects;
 
+import com.uniqueAuction.web.trade.request.SaleRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
@@ -12,9 +13,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.uniqueAuction.domain.trade.service.SaleService;
 import com.uniqueAuction.web.response.CommonResponse;
-import com.uniqueAuction.web.trade.request.SaleBidRequest;
-import com.uniqueAuction.web.trade.request.SaleNowRequest;
-
 import lombok.RequiredArgsConstructor;
 
 @RestController
@@ -23,24 +21,14 @@ public class SaleController {
 
 	private final SaleService saleService;
 
-	@PostMapping("/saleBid")
+	@PostMapping("/sale")
 	@ResponseStatus(HttpStatus.CREATED)
-	public CommonResponse saveSaleBid(@RequestBody @Validated SaleBidRequest saleBidRequest, BindingResult result) {
-		if (result.hasErrors()) {
-			throw new RuntimeException(Objects.requireNonNull(result.getFieldError()).getDefaultMessage());
-		}
-		saleService.saveSaleBid(saleBidRequest.toEntity());
-		return CommonResponse.success();
-	}
-
-	@PostMapping("/saleNow")
-	@ResponseStatus(HttpStatus.CREATED)
-	public CommonResponse saveSaleNow(@RequestBody @Validated SaleNowRequest saleNowRequest,
+	public CommonResponse<?> saveSaleNow(@RequestBody @Validated SaleRequest saleRequest,
 		BindingResult result) {
 		if (result.hasErrors()) {
 			throw new RuntimeException(Objects.requireNonNull(result.getFieldError()).getDefaultMessage());
 		}
-		saleService.saveSaleNow(saleNowRequest.toEntity());
+		saleService.saveSale(saleRequest.toEntity());
 		return CommonResponse.success();
 	}
 }
