+++ src/main/java/com/uniqueAuction/web/trade/request/SaleRequest.java
@@ -2,12 +2,12 @@ package com.uniqueAuction.web.trade.request;
 
 import javax.validation.constraints.NotBlank;
 
-import com.uniqueAuction.domain.trade.entity.Purchase;
+import com.uniqueAuction.domain.trade.entity.Sale;
 
 import lombok.Getter;
 
 @Getter
-public class PurchaseBidRequest {
+public class SaleRequest {
 
 	@NotBlank(message = "유저ID는 필수값입니다.")
 	private Long userId;
@@ -21,16 +21,16 @@ public class PurchaseBidRequest {
 	@NotBlank(message = "입찰가격은 필수값입니다.")
 	private String bidPrice;
 
-	@NotBlank(message = "배송주소는 필수값입니다.")
-	private String shippingAddress;
+	@NotBlank(message = "반송주소는 필수값입니다.")
+	private String returnAddress;
 
-	public Purchase toEntity() {
-		return Purchase.builder()
+	public Sale toEntity() {
+		return Sale.builder()
 			.userId(this.userId)
 			.modelNumber(this.modelNumber)
 			.bidPrice(this.bidPrice)
 			.productSize(this.productSize)
-			.shippingAddress(this.shippingAddress)
+			.returnAddress(this.returnAddress)
 			.build();
 	}
 }
